+++ pkgs/ffigen/lib/src/code_generator/func_type.dart
@@ -69,6 +69,12 @@ class FunctionType extends Type {
       _getTypeImpl(writeArgumentNames, (Type t) => t.getDartType(w));
 
   @override
+  String getNativeType({String varName = ''}) {
+    final arg = dartTypeParameters.map<String>((p) => p.type.getNativeType());
+    return '${returnType.getNativeType()} (*$varName)(${arg.join(', ')})';
+  }
+
+  @override
   bool get sameFfiDartAndCType =>
       returnType.sameFfiDartAndCType &&
       dartTypeParameters.every((p) => p.type.sameFfiDartAndCType);
@@ -140,6 +146,10 @@ class NativeFunc extends Type {
 
   @override
   String getFfiDartType(Writer w) => getCType(w);
+
+  @override
+  String getNativeType({String varName = ''}) =>
+      _type.getNativeType(varName: varName);
 
   @override
   bool get sameFfiDartAndCType => true;