+++ pkgs/ffigen/lib/src/code_generator/typealias.dart
@@ -131,6 +131,10 @@ class Typealias extends BindingType {
   String getCType(Writer w) => name;
 
   @override
+  String getNativeType({String varName = ''}) =>
+      type.getNativeType(varName: varName);
+
+  @override
   String getFfiDartType(Writer w) {
     if (_ffiDartAliasName != null) {
       return _ffiDartAliasName!;
@@ -184,6 +188,9 @@ class Typealias extends BindingType {
       );
 
   @override
+  String? generateRetain(String value) => type.generateRetain(value);
+
+  @override
   String cacheKey() => type.cacheKey();
 
   @override
@@ -230,4 +237,7 @@ class ObjCInstanceType extends Typealias {
             )
           : ObjCInterface.generateConstructor(
               objCEnclosingClass, value, objCRetain);
+
+  @override
+  String getNativeType({String varName = ''}) => 'id $varName';
 }